requirements.txt
discord.py-self==1.9.2
colorama
requests
pywin32
pycryptodome
